Move dataclass definitions to separate init_dynamodb.py
